add a lighter duration function
