+++ .gitignore
@@ -1,3 +1,3 @@
 .pnpm-debug.log
 dumpster.js
-testing/test.js
+testing
